Move integration tests to their own class
